pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "expense-tracker-api"
version = "0.1.0"
description = "An API for an expense tracker application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.22",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
]
